pyproject.toml
[tool.poetry]
name = "django-task"
version = "0.1.0"
description = ""
authors = ["saatarko <saatar@tut.by>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
